refactor: Minor refactoring
- form_validator_locale.dart renamed to locale.dart
- imports optimized
