+++ lib/form_validator.dart
@@ -1,5 +1,5 @@
 library form_validator;
 
-export 'src/form_validator_locale.dart';
+export 'src/locale.dart';
 export 'src/validator_builder.dart'
     show ValidationBuilder, StringValidationCallback;
