+++ lib/src/i18n/az.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocaleAz implements FormValidatorLocale {
   const LocaleAz();
