+++ lib/src/i18n/de.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocaleDe implements FormValidatorLocale {
   const LocaleDe();
