+++ lib/src/i18n/es.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocaleEs implements FormValidatorLocale {
   const LocaleEs();
