+++ lib/src/i18n/id.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocaleId implements FormValidatorLocale {
   const LocaleId();
