+++ lib/src/i18n/pt_br.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocalePtBr implements FormValidatorLocale {
   const LocalePtBr();
