+++ lib/src/i18n/zh_cn.dart
@@ -1,4 +1,4 @@
-import '../form_validator_locale.dart';
+import '../locale.dart';
 
 class LocaleZhCN implements FormValidatorLocale {
   const LocaleZhCN();
